+++ CorePluginFramework/src/au/gov/asd/tac/constellation/plugins/templates/CascadingQueryPlugin.java
@@ -65,11 +65,6 @@ import java.util.logging.Logger;
  *
  * @author sirius
  */
-//@Messages({
-//    "# {0} - graph",
-//    "# {1} - name",
-//    "MSG_Query_Failed=Action failed: {0}; {1}"
-//})
 public abstract class CascadingQueryPlugin extends AbstractPlugin {
 
     private static final Logger LOGGER = Logger.getLogger(CascadingQueryPlugin.class.getName());