+++ CorePluginFramework/src/au/gov/asd/tac/constellation/plugins/templates/SimpleReadPlugin.java
@@ -23,12 +23,7 @@ import au.gov.asd.tac.constellation.plugins.AbstractPlugin;
 import au.gov.asd.tac.constellation.plugins.PluginException;
 import au.gov.asd.tac.constellation.plugins.PluginGraphs;
 import au.gov.asd.tac.constellation.plugins.PluginInteraction;
-import au.gov.asd.tac.constellation.plugins.PluginNotificationLevel;
 import au.gov.asd.tac.constellation.plugins.parameters.PluginParameters;
-import au.gov.asd.tac.constellation.utilities.text.SeparatorConstants;
-import java.util.logging.Level;
-import java.util.logging.Logger;
-import org.openide.util.NbBundle;
 
 /**
  * A plugin template for plugins that only require read access to the graph.
@@ -44,14 +39,7 @@ import org.openide.util.NbBundle;
  *
  * @author sirius
  */
-@NbBundle.Messages({
-    "# {0} - graph",
-    "# {1} - name",
-    "MSG_Read_Failed=Action failed: {0}; {1}"
-})
 public abstract class SimpleReadPlugin extends AbstractPlugin {
-
-    private static final Logger LOGGER = Logger.getLogger(SimpleReadPlugin.class.getName());
 
     public SimpleReadPlugin() {
         // This constructor is intentionally left blank.
@@ -62,95 +50,52 @@ public abstract class SimpleReadPlugin extends AbstractPlugin {
     }
 
     @Override
-    public final void run(final PluginGraphs graphs, final PluginInteraction interaction, final PluginParameters parameters) throws InterruptedException, PluginException, RuntimeException {
+    public final void run(final PluginGraphs graphs, final PluginInteraction interaction, final PluginParameters parameters) throws InterruptedException, PluginException {
         final Graph graph = graphs.getGraph();
 
         // Make the graph appear busy
         interaction.setBusy(graph.getId(), true);
         try {
-
             // Make the progress bar appear nondeterminent
             interaction.setProgress(0, 0, "Waiting...", true);
+
             try {
-
                 ReadableGraph readableGraph = graph.getReadableGraph();
+
                 try {
-
                     interaction.setProgress(0, 0, "Working...", true);
-                    try {
-                        read(readableGraph, interaction, parameters);
-                    } catch (final InterruptedException e) {
-                        // Notify the user that the plugin was interrupted and throw the exception back to the caller for generic handling.
-                        interaction.notify(PluginNotificationLevel.INFO, "Plugin cancelled: " + graphs.getGraph() + SeparatorConstants.SEMICOLON + " " + getName());
-                        throw e;
-                    } catch (final PluginException e) {
-                        // Logging the PluginException and throwing the exception back to the caller for generic handling.
-                        final String msg0 = String.format("PluginException caught in %s.run()", SimpleReadPlugin.class.getName());
-                        final String msg = Bundle.MSG_Read_Failed(graph, getName());
-                        LOGGER.log(Level.SEVERE, "{0}" + SeparatorConstants.SEMICOLON + SeparatorConstants.NEWLINE + "{1}" + SeparatorConstants.NEWLINE + "{2}", new Object[]{msg0, msg, e.getMessage()});
-                        throw e;
-                    } catch (final RuntimeException e) {
-                        // Notify the user that there was a RuntimeException and throw the exception back to the caller for generic handling.
-                        final String msg0 = String.format("Unexpected non-plugin exception caught in %s.run()", SimpleReadPlugin.class.getName());
-                        final String msg = Bundle.MSG_Read_Failed(graph, getName());
-                        interaction.notify(PluginNotificationLevel.ERROR, msg0 + SeparatorConstants.SEMICOLON + SeparatorConstants.NEWLINE + msg + SeparatorConstants.NEWLINE + e.getMessage());
-                        throw e;
-                    }
-
+                    read(readableGraph, interaction, parameters);
                 } finally {
                     readableGraph.release();
                 }
-
             } finally {
                 interaction.setProgress(2, 1, "Finished", true);
             }
-
         } finally {
             interaction.setBusy(graph.getId(), false);
         }
     }
 
     @Override
-    public void run(final GraphReadMethods graph, final PluginInteraction interaction, final PluginParameters parameters) throws InterruptedException, PluginException, RuntimeException {
+    public void run(final GraphReadMethods graph, final PluginInteraction interaction, final PluginParameters parameters) throws InterruptedException, PluginException {
         // Make the graph appear busy
         interaction.setBusy(graph.getId(), true);
         try {
-
             // Make the progress bar appear nondeterminent
             interaction.setProgress(0, 0, "Working...", true);
+
             try {
-
-                try {
-                    read(graph, interaction, parameters);
-                } catch (final InterruptedException e) {
-                    // Notify the user that the plugin was interrupted and throw the exception back to the caller for generic handling.
-                    interaction.notify(PluginNotificationLevel.INFO, "Plugin cancelled: " + graph + SeparatorConstants.COLON + " " + getName());
-                    throw e;
-                } catch (final PluginException e) {
-                    // Logging the PluginException and throwing the exception back to the caller for generic handling.
-                    final String msg0 = String.format("PluginException caught in %s.run()", SimpleReadPlugin.class.getName());
-                    final String msg = Bundle.MSG_Read_Failed(graph, getName());
-                    LOGGER.log(Level.SEVERE, "{0}" + SeparatorConstants.SEMICOLON + SeparatorConstants.NEWLINE + "{1}" + SeparatorConstants.NEWLINE + "{2}", new Object[]{msg0, msg, e.getMessage()});
-                    throw e;
-                } catch (final RuntimeException e) {
-                    // Notify the user that there was a RuntimeException and throw the exception back to the caller for generic handling.
-                    final String msg0 = String.format("Unexpected non-plugin exception caught in %s.run()", SimpleReadPlugin.class.getName());
-                    final String msg = Bundle.MSG_Read_Failed(graph, getName());
-                    interaction.notify(PluginNotificationLevel.ERROR, msg0 + SeparatorConstants.SEMICOLON + SeparatorConstants.NEWLINE + msg + SeparatorConstants.NEWLINE + e.getMessage());
-                    throw e;
-                }
-
+                read(graph, interaction, parameters);
             } finally {
                 interaction.setProgress(2, 1, "Finished", true);
             }
-
         } finally {
             interaction.setBusy(graph.getId(), false);
         }
     }
 
     @Override
-    public final void run(final GraphWriteMethods graph, final PluginInteraction interaction, final PluginParameters parameters) throws InterruptedException, PluginException, RuntimeException {
+    public final void run(final GraphWriteMethods graph, final PluginInteraction interaction, final PluginParameters parameters) throws InterruptedException, PluginException {
         run((GraphReadMethods) graph, interaction, parameters);
     }
 
@@ -166,8 +111,6 @@ public abstract class SimpleReadPlugin extends AbstractPlugin {
      * @throws InterruptedException if the plugin execution is canceled.
      * @throws PluginException if an anticipated error occurs during plugin
      * execution.
-     * @throws RuntimeException if an unexpected error occurs during the plugin
-     * execution.
      */
-    protected abstract void read(GraphReadMethods graph, PluginInteraction interaction, PluginParameters parameters) throws InterruptedException, PluginException, RuntimeException;
+    protected abstract void read(GraphReadMethods graph, PluginInteraction interaction, PluginParameters parameters) throws InterruptedException, PluginException;
 }